agent 02/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
typing
requests
chromadb
sentence-transformers